Work on Generator Pipeline. WIP

+++ src/diffusers/pipelines/wuerstchen/modules.py
@@ -328,7 +328,7 @@ class DiffNeXt(ModelMixin, ConfigMixin):
         emb = torch.cat([emb.sin(), emb.cos()], dim=1)
         if self.c_r % 2 == 1:  # zero pad
             emb = nn.functional.pad(emb, (0, 1), mode="constant")
-        return emb
+        return emb.to(dtype=r.dtype)
 
     def gen_c_embeddings(self, clip):
         clip = self.clip_mapper(clip)
